refactor(shielded): relocate createShieldedOutputs input types to shielded/types
`ShieldedOutputProposal` (the input shape of `createShieldedOutputs`)
was defined locally in `creation.ts` and not exported, so external
callers — including any consumer added in later PRs — couldn't
construct the input array without redefining the shape. The other two
input-side types in `creation.ts` (`ShieldedInputBlinding`,
`InputGeneratorInfo`) were exported but lived alongside it for no
particular reason.

Move all three to `src/shielded/types.ts`, alongside `IDataShieldedOutput`
(the return-side type that was relocated in an earlier commit on this
PR). All four createShieldedOutputs I/O types now live next to each
other in the shielded types file under a "createShieldedOutputs I/O"
section header; `creation.ts` imports them from `./types`.

No behavior change. All 12 existing creation tests still pass.
